refactor(purchaser): split handler.py into focused modules
handler.py was 1009 lines mixing orchestration, coverage math, guards
(cooldown + spike), purchase execution, and email formatting. Split into:

- handler.py (142 lines): pipeline orchestration only
- coverage_calc.py (161 lines): current coverage, expiring-plan handling
- guards.py (209 lines): apply_spike_guard + apply_purchase_cooldown and
  their SNS notifications (previously private functions)
- purchase_execution.py (282 lines): process_purchase_messages,
  execute_purchase, summary email builder

No behavior change. Module name is coverage_calc (not coverage) to avoid
shadowing the coverage package used by pytest-cov.

All 31 purchaser tests continue to pass via the existing handler.handler()
entry point — no internal symbols are imported by tests.
